requirements.txt
flask
python-dotenv
openai
requests
supabase
flask-cors
gunicorn
